+++ EssentialApp/EssentialAppTests/FeedLoaderCacheDecorator.swift
@@ -20,7 +20,7 @@ final class FeedLoaderCacheDecorator: FeedLoader {
     }
 }
 
-final class FeedLoaderCacheDecoratorTests: XCTestCase {
+final class FeedLoaderCacheDecoratorTests: XCTestCase, FeedLoaderTestCase {
     func test_load_deliversFeedOnLoaderSuccess() {
         let feed = uniqueFeed().model
         let loader = FeedLoaderStub(result: .success([feed]))
@@ -38,25 +38,4 @@ final class FeedLoaderCacheDecoratorTests: XCTestCase {
     }
     
     // MARK: - Helpers
-    func expect(
-        _ sut: FeedLoader,
-    toFinishWith expectedResult: FeedLoader.Result,
-    file: StaticString = #file,
-    line: UInt = #line
-    ) {
-        let expectation = XCTestExpectation(description: "Wait for load")
-        sut.load { actualResult in
-            switch (actualResult, expectedResult) {
-                case (.success(let actual), .success(let expected)):
-                    XCTAssertEqual(actual, expected, file: file, line: line)
-            case  (.failure(let actual as NSError), .failure(let expected as NSError)):
-                XCTAssertEqual(actual, expected, file: file, line: line)
-                default:
-                    XCTFail("Expected \(expectedResult), got \(actualResult) instead", file: file, line: line)
-            }
-            expectation.fulfill()
-        }
-        
-        wait(for: [expectation], timeout: 1.0)
-    }
 }
